requirements.txt
Flask
Flask-Bcrypt
bcrypt
bs4
django
python-dotenv
psycopg2-binary
vercel
gunicorn
requests
flask-socketio 
eventlet
gevent
gevent-websocket
pytz
